backend/utils/transcribe.py
import os
import requests
from dotenv import load_dotenv

load_dotenv()

async def transcribe_audio(file_path: str) -> str:
    """
    Transcribe audio using OpenAI Whisper API (compatible with Groq setup)
    Note: Groq doesn't have speech-to-text, so we use OpenAI Whisper
    """
    try:
        # Since Groq doesn't offer speech-to-text, we'll use OpenAI Whisper
        # You can get OpenAI API key for this specific purpose
        openai_api_key = os.getenv("OPENAI_API_KEY")  # Optional: for Whisper
        
        if openai_api_key:
            # Use OpenAI Whisper if available
            from openai import AsyncOpenAI
            client = AsyncOpenAI(api_key=openai_api_key)
            
            with open(file_path, "rb") as audio_file:
                transcript = await client.audio.transcriptions.create(
                    model="whisper-1",
                    file=audio_file,
                    response_format="text"
                )
            return transcript
        else:
            # Fallback: Simulate transcription for demo
            print("No OpenAI API key found. Using simulated transcription.")
            return simulate_transcription()
            
    except Exception as e:
        print(f"Transcription error: {str(e)}")
        # Fallback to simulated transcription
        return simulate_transcription()

def simulate_transcription() -> str:
    """
    Simulated transcription for demo purposes
    Replace this with actual Whisper API call in production
    """
    return """
    John: Okay team, let's start our weekly project meeting. First, Sarah, can you update us on the frontend development?
    Sarah: Sure, we've completed the user authentication module and are now working on the dashboard interface. We're about 70% done.
    Mike: On the backend, I've optimized the database queries. We're seeing 40% faster response times now.
    John: That's great progress. Key decision - we'll deploy to staging environment this Friday.
    Sarah: I'll make sure the frontend is ready by Thursday. Action item - Mike will prepare the deployment checklist.
    Mike: I'll also document the API changes for the mobile team. Deadline: end of this week.
    John: Perfect. Let's aim for production deployment next Wednesday. Any concerns?
    Sarah: None from my side. The design team has approved all mockups.
    Mike: Backend is stable and ready for integration.
    John: Excellent. Meeting adjourned.
    """